fix(evm): narrow the program counter for the debug tracer
`DebugTracer.TryWait` assigns the counter into `VmState.ProgramCounter`,
which stays `int`. The call site is behind `#if DEBUG`, so the release
build did not catch it.

+++ src/Nethermind/Nethermind.Evm/Tracing/Debugger/DebugTracer.cs
@@ -91,7 +91,7 @@ public class DebugTracer<TGasPolicy>(ITxTracer tracer) : ITxTracer, ITxTracerWra
 
         lock (_lock)
         {
-            vmState.ProgramCounter = programCounter;
+            vmState.ProgramCounter = (int)programCounter;
             vmState.Gas = gas;
             vmState.DataStackHead = stackHead;
             CurrentState = vmState;
